requirements.txt
# 基于EEG的吸烟成瘾评估系统 - 依赖包列表
# 此文件列出了项目运行所需的所有Python包及其版本

# ===== 脑电数据处理核心库 =====
mne>=1.5.0                    # MNE-Python: 脑电数据处理和分析的主要库
mne-bids>=0.12                # MNE-BIDS: 支持BIDS格式的脑电数据
eeglabio>=0.1.0               # EEGLAB格式数据支持

# ===== 科学计算和数值处理 =====
numpy>=1.21.0                 # 数值计算基础库，提供多维数组和数学函数
scipy>=1.7.0                  # 科学计算库，提供信号处理、优化等算法
pandas>=1.3.0                 # 数据处理和分析库，提供DataFrame结构
scikit-learn>=1.0.0           # 机器学习库，提供分类、回归、聚类等算法

# ===== 信号处理和特征提取 =====
pywt>=1.4.0                   # PyWavelets: 小波变换库
spectral>=0.22.0              # 频谱分析库
antropy>=0.1.4                # 熵计算库，用于复杂度分析

# ===== 可视化和报告 =====
matplotlib>=3.5.0             # 基础绘图库
seaborn>=0.11.0               # 基于matplotlib的统计可视化库
plotly>=5.0.0                 # 交互式可视化库
dash>=2.0.0                   # Web应用框架，用于创建交互式仪表板
dash-bootstrap-components>=1.0 # Dash的Bootstrap组件

# ===== 配置和实验管理 =====
hydra-core>=1.2.0             # 配置管理库，支持YAML配置和命令行参数
omegaconf>=2.2.0              # 配置对象库，Hydra的配置后端

# ===== 数据存储和序列化 =====
h5py>=3.6.0                   # HDF5文件格式支持
joblib>=1.1.0                 # 并行计算和模型保存库
pickle5>=0.0.11               # 对象序列化库

# ===== 其他工具 =====
tqdm>=4.62.0                  # 进度条显示库
rich>=12.0.0                  # 终端美化库
click>=8.0.0                  # 命令行接口库

# ===== 开发依赖（可选） =====
# 以下包用于代码质量检查、测试和开发，可通过 pip install -e ".[dev]" 安装

# 代码质量检查
ruff>=0.0.270                 # 快速Python代码检查器
mypy>=0.991                   # 静态类型检查器
black>=22.0.0                 # 代码格式化工具
isort>=5.10.0                 # import语句排序工具

# 测试框架
pytest>=7.0.0                 # 测试框架
pytest-cov>=4.0.0             # 测试覆盖率插件
pytest-mock>=3.8.0            # 测试mock插件

# 预提交钩子
pre-commit>=2.20.0            # Git预提交钩子管理

# 文档生成
sphinx>=5.0.0                 # 文档生成器
sphinx-rtd-theme>=1.0.0       # ReadTheDocs主题

# 性能分析
memory-profiler>=0.60.0       # 内存使用分析工具
line-profiler>=3.5.0          # 逐行性能分析工具

# ===== 额外依赖 =====
networkx>=2.8.0               # 复杂网络分析库，用于功能连接分析
PyYAML>=6.0                   # YAML文件解析库

